feat(phase-5.5): main-batch driver + reliability hardening
Phase 5.5 main-batch infrastructure that brought 8000 cells x 2 models to
completion on the remote server. Bundles the script and several runner-side
reliability fixes that surfaced during the long run.

Driver:
- scripts/_main_batch.py: 8000-cell x 2-model orchestrator with
  --chunk-size, --limit, --max-cost (per-process budget abort), --model
  (single-model parallel split for vendor-side rate-limit control), --rpm
  (per-process pacing); per-call tqdm bar; per-model output paths
  scores_{glm-5,glm-4.5}.parquet to avoid concurrent-write race when
  running two procs in parallel.
- scripts/run_main_batch.sh: bash wrapper (sanity checks + proxy unset
  for direct China-mainland Zhipu; forwards args).

Client / runner reliability:
- ZhipuClient: explicit per-request HTTP timeout (default 60s) plus
  max_output_tokens (default 600) -- prior runs hung indefinitely on
  stalled responses and burned budget on long reasoning emissions.
- BatchRunner._load_existing_scores: dedup only on rows with a
  non-null hiring_score, so failed cells get retried on resume instead
  of being treated as final (root cause of the multi-day rate-limit
  freezes earlier in the run).
- BatchRunner._score_one: tenacity backoff bumped from max=30s to
  max=120s to absorb Zhipu account-level rate-limit (1302) recovery.
- BatchRunner.run_with_prompts: optional progress_cb for per-call
  progress reporting (consumed by _main_batch.py's tqdm bar).

Schema:
- ScoringResult.hiring_score: int -> float, aligning with the
  post-v0.5b float schema in utils/prompts.py.

Prompt:
- SCORING_SYSTEM_PROMPT + JSON schema rationale field tightened to
  one or two sentences <= 200 chars naming the single most decisive
  evidence, to cap output-token cost on glm-5 reasoning.

+++ src/llm_audit/scoring/base.py
@@ -25,7 +25,7 @@ class ScoringResult:
     cell_id: int
     model_id: str
     provider: str
-    hiring_score: int
+    hiring_score: float
     rationale: str
     latency_ms: int
     tokens_in: int
